pyproject.toml
[tool.poetry]
name = "h8-unet-activefire"
version = "0.1.0"
description = "Starting to prototype the active fire detection using U-Net but mostly will continue with the same repo"
authors = ["Jayendra"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
pandas = "^2.2.0"
geopandas = "^0.14.3"
h5py = "^3.10.0"
rasterio = "^1.3.9"
tqdm = "^4.66.2"
torchgeo = "^0.5.1"
torchvision = "^0.17.0"
scikit-learn = "^1.4.1.post1"
scipy = "^1.12.0"
scikit-image = "^0.22.0"
matplotlib = "^3.8.3"
opencv-python-headless = "^4.9.0.80"
kornia = "^0.7.1"
pyproj = "^3.6.1"
tensorboard = "^2.16.2"
black = "^24.2.0"
boto3 = "^1.34.104"
satpy = "^0.48.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"


[tool.black]
line-length = 88
target-version = ['py38']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
